增加系统环境判断，避免重复安装

+++ install/MySQL.sh
@@ -2,7 +2,7 @@
 PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
 echo -e "# ******************************************************"
 echo -e "#                                                      "*
-echo -e "# *脚本更新时间：2023年10月20日                         "*
+echo -e "# *脚本更新时间：2023年10月21日                         "*
 echo -e "#                                                      "*
 echo -e "# *脚本支持CentOS/Ubuntu/Debian                        "* 
 echo -e "#                                                      "*
@@ -10,7 +10,8 @@ echo -e "# *可按Ctrl+Z取消安装                                  "*
 echo -e "#                                                      "*
 echo -e "# ******************************************************"
 echo -e "                                                       "
-
+# 获取内网IP地址
+IP=$(ip addr | grep -E -o '[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}' | grep -E -v "^127\.|^255\.|^0\." | head -n 1)
 # 检测系统类型
 if [[ -f /etc/redhat-release ]]; then
 # CentOS
@@ -18,55 +19,143 @@ echo -e "# 检测到系统为CentOS，仅支持MySQL5.7"
 # 关闭并禁用防火墙
 systemctl stop firewalld
 systemctl disable firewalld
-# 检测MySQL和Redis安装包
+# 检测MySQL、Redis、Docker安装包
 while true; do
-    if [ ! -f mysql-5.7*.tar ]; then
-        echo "需要的 mysql-5.7.tar 文件不存在，是否通过网络下载？(Y/N)"
+    if ! command -v mysql &> /dev/null; then
+        echo "未检测到MySQL环境，是否通过网络下载并安装？(Y/N)"
         read -r download_mysql
         if [ "$download_mysql" = "Y" ] || [ "$download_mysql" = "y" ]; then
+            echo "正在下载 MySQL 安装包..."
             yum install -y wget && wget -c -N --no-check-certificate https://cdn.mysql.com//Downloads/MySQL-5.7/mysql-5.7.43-1.el7.x86_64.rpm-bundle.tar
-            break  # 如果下载完成，退出循环
         else
-            echo "请将 mysql-5.7.tar 文件放置到当前目录下，然后按回车键继续..."
+            echo "请将 mysql.tar 文件放置到当前目录下，然后按回车键继续..."
             read  # 这里等待用户按回车键
+            if ls mysql*.tar 1> /dev/null 2>&1; then
+                echo "检测到MySQL安装包已存在。"                
+                # 移除任何已经安装的 MySQL 或者 MariaDB
+                rpm -e `rpm -qa | grep -i mysql`
+                rpm -e --nodeps `rpm -qa | grep -i mariadb`
+                # 解压Mysql5.7安装包
+                tar -xvf mysql-5.7*.tar 
+                # 安装Mysql5.7
+                rpm -ivh mysql-community-common-5.7*.rpm
+                rpm -ivh mysql-community-libs-5.7*.rpm
+                rpm -ivh mysql-community-client-5.7*.rpm
+                rpm -ivh --nodeps mysql-community-server-5.7*.rpm
+                # 启动Mysql5.7
+                systemctl start mysqld
+                break  # 如果下载完成，退出循环
+            fi
         fi
     else
-        break  # 如果文件存在，退出循环
+        echo "检测到已有MySQL环境，跳过安装"
+        break  # 如果MySQL已安装，退出循环
     fi
 done
 while true; do
-    if [ ! -f redis*.rpm  ]; then
-        echo "需要的 redis.rpm文 文件不存在，是否通过网络下载？(Y/N)"
+    if ! command -v redis-server &> /dev/null; then
+        echo "未检测到Redis环境，是否通过网络下载并安装？(Y/N)"
         read -r download_redis
         if [ "$download_redis" = "Y" ] || [ "$download_redis" = "y" ]; then
+            echo "正在下载 Redis 安装包..."
             yum install -y wget && wget -c -N --no-check-certificate https://rpms.remirepo.net/enterprise/7/remi/x86_64/redis-7.2.2-1.el7.remi.x86_64.rpm
             break  # 如果下载完成，退出循环
         else
-            echo "请将 redis.rpm文 文件放置到当前目录下，然后按回车键继续..."
+            echo "请将 redis.rpm 文件放置到当前目录下，然后按回车键继续..."
             read  # 这里等待用户按回车键
-        fi
+            if ls redis*.rpm 1> /dev/null 2>&1; then
+                echo "检测到Redis安装包已存在。"
+                # 安装Redis
+                rpm -ivh redis*.rpm
+                break  # 如果下载完成，退出循环
+            fi
+        fi       
     else
-        break  # 如果文件存在，退出循环
+        echo "检测到已有Redis环境，跳过安装"
+        break  # 如果Redis已安装，退出循环
     fi
 done
-# 移除任何已经安装的 MySQL 或者 MariaDB
-rpm -e `rpm -qa | grep -i mysql`
-rpm -e --nodeps `rpm -qa | grep -i mariadb`
+while true; do
+    if ! command -v docker &> /dev/null; then
+        echo "未检测到Docker环境，是否通过网络下载并安装？(Y/N)"
+        read -r download_docker
+         if [ "$download_docker" = "Y" ] || [ "$download_docker" = "y" ]; then
+            curl -sSL https://get.docker.com/ | sh 
+            systemctl enable docker && systemctl start docker            
+            break  # 如果下载完成，退出循环
+        else
+            echo "请将 docker.tgz 文件放置到当前目录下，然后按回车键继续..."
+            read  # 这里等待用户按回车键
+            if ls docker*.tgz 1> /dev/null 2>&1; then
+                echo "检测到Docker安装包已存在。"
+# 安装docker
+tar -zxvf docker*.tgz
+cp docker/* /usr/bin/
 
-# 解压Mysql5.7安装包
-tar -xvf mysql-5.7*.tar 
+# 写入docker.service的内容
+cat <<EOL > /etc/systemd/system/docker.service
+[Unit]
+Description=Docker Application Container Engine
+Documentation=https://docs.docker.com
+After=network-online.target firewalld.service
+Wants=network-online.target
 
-# 安装Mysql5.7
-rpm -ivh mysql-community-common-5.7*.rpm
-rpm -ivh mysql-community-libs-5.7*.rpm
-rpm -ivh mysql-community-client-5.7*.rpm
-rpm -ivh --nodeps mysql-community-server-5.7*.rpm
+[Service]
+Type=notify
+# 默认情况下不使用systemd管理cgroup，因为委托问题仍然存在
+# 而且systemd目前不支持运行docker容器所需的cgroup特性集
+ExecStart=/usr/bin/dockerd --selinux-enabled=false --insecure-registry=172.16.40.210
+ExecReload=/bin/kill -s HUP \$MAINPID
+# 非零的Limit*s会导致由于内核中的会计开销而出现性能问题
+# 我们建议使用cgroups来进行容器的本地账户管理
+LimitNOFILE=infinity
+LimitNPROC=infinity
+LimitCORE=infinity
 
-# 启动Mysql5.7
-systemctl start mysqld
+# 如果你的systemd版本支持，取消注释TasksMax
+# 仅systemd 226及以上版本支持此选项
+#TasksMax=infinity
 
-# 安装Redis
-rpm -ivh redis*.rpm
+TimeoutStartSec=0
+# 设置委派为yes，以便systemd不会重置docker容器的cgroup
+Delegate=yes
+# 仅终止docker进程，而不终止cgroup中的所有进程
+KillMode=process
+# 如果docker进程异常退出，则重启docker进程
+Restart=on-failure
+StartLimitBurst=3
+StartLimitInterval=60s
+
+[Install]
+WantedBy=multi-user.target
+EOL
+
+# 修改docker.service文件中的ExecStart行
+sed -i "s|ExecStart=/usr/bin/dockerd.*|ExecStart=/usr/bin/dockerd --selinux-enabled=false --insecure-registry=$IP|" /etc/systemd/system/docker.service
+
+# 限制docker日志大小
+mkdir /etc/docker
+cat >/etc/docker/daemon.json<<EOF
+{
+"log-driver": "json-file",
+"log-opts": {"max-size":"20m", "max-file":"2"}
+}
+EOF
+
+# 重新加载systemd并启用新的Docker服务
+systemctl daemon-reload
+systemctl enable docker.service
+
+# 使用新的服务启动Docker
+systemctl start docker.service
+                break  # 如果检测存在，退出循环
+            fi
+        fi
+    else
+        echo "检测到已有Docker环境，跳过安装"
+        break  # 如果Docker已安装，退出循环
+    fi
+done
 
 # 提取临时密码
 temp_password=$(grep 'temporary password' /var/log/mysqld.log | awk '{print $NF}')
@@ -94,8 +183,8 @@ echo "Mysql密码已修改为${new_password}"
 
 # 创建同步所需的MYSQL用户
 read -ep "请输入MySQL主服务器的地址: " master_host
-read -ep "请输入同步用的MySQL用户名: " master_user
-read -ep "请输入同步用的MySQL密码: " master_password
+read -ep "请输入同步用的MySQL用户名（主从保持一致）: " master_user
+read -ep "请输入同步用的MySQL密码（主从保持一致）: " master_password
 read -ep "请输入MySQL服务器的ID（输入数字，注意不能重复，默认1为主服务器）: " mysql_id
 
 cat  >> /etc/my.cnf <<EOF 
@@ -166,33 +255,40 @@ echo -e "# 检测到系统为Ubuntu，仅支持MySQL5.7"
 sudo ufw disable
 # 检测MySQL安装包
 while true; do
-    if [ ! -f mysql*.tar ]; then
-        echo "需要的 mysql.tar 文件不存在，是否通过网络下载？(Y/N)"
+    if ! command -v mysql &> /dev/null; then
+        echo "未检测到MySQL环境，是否通过网络下载并安装？(Y/N)"
         read -r download_mysql
         if [ "$download_mysql" = "Y" ] || [ "$download_mysql" = "y" ]; then
-            wget -c -N --no-check-certificate https://downloads.mysql.com/archives/get/p/23/file/mysql-server_5.7.42-1ubuntu18.04_amd64.deb-bundle.tar
-            break  # 如果下载完成，退出循环
+            echo "正在下载 MySQL 安装包..."
+            yum install -y wget && wget -c -N --no-check-certificate https://downloads.mysql.com/archives/get/p/23/file/mysql-server_5.7.42-1ubuntu18.04_amd64.deb-bundle.tar
         else
             echo "请将 mysql.tar 文件放置到当前目录下，然后按回车键继续..."
             read  # 这里等待用户按回车键
+            if ls mysql*.tar 1> /dev/null 2>&1; then
+                echo "检测到MySQL安装包已存在。"
+                # 解压MySQL安装包
+                tar -xvf mysql*.tar
+                # 安装MySQL
+                apt install -y libaio1 libtinfo5 libmecab2 redis
+                dpkg -i mysql-common*.deb
+                dpkg-preconfigure mysql-community-server*.deb 
+                dpkg -i libmysqlclient20*.deb
+                dpkg -i libmysqlclient-dev*.deb
+                dpkg -i libmysqld-dev*.deb
+                dpkg -i mysql-community-client*.deb
+                dpkg -i mysql-client*.deb
+                dpkg -i mysql-community-server*.deb
+                dpkg -i mysql-server*.deb
+                break  # 如果下载完成，退出循环
+            fi
         fi
     else
-        break  # 如果文件存在，退出循环
+        echo "检测到已有MySQL环境，跳过安装"
+        break  # 如果MySQL已安装，退出循环
     fi
 done
-# 解压MySQL安装包
-tar -xvf mysql*.tar
-# 安装MySQL
-apt install -y libaio1 libtinfo5 libmecab2 redis
-dpkg -i mysql-common*.deb
-dpkg-preconfigure mysql-community-server*.deb 
-dpkg -i libmysqlclient20*.deb
-dpkg -i libmysqlclient-dev*.deb
-dpkg -i libmysqld-dev*.deb
-dpkg -i mysql-community-client*.deb
-dpkg -i mysql-client*.deb
-dpkg -i mysql-community-server*.deb
-dpkg -i mysql-server*.deb
+# 安装Redis
+apt install -y redis
 # 获取MySQL密码
 read -ep "请输入主MySQL的root密码: " root_password
 # 修改密码并允许所有IP访问
@@ -206,8 +302,8 @@ echo "Redis密码已修改为${root_password}"
 mysql -uroot -p"$root_password" -e "GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY '${root_password}'; FLUSH PRIVILEGES;"
 # 创建同步所需的MYSQL用户
 read -ep "请输入MySQL主服务器的地址: " master_host
-read -ep "请输入同步用的MySQL用户名: " master_user
-read -ep "请输入同步用的MySQL密码: " master_password
+read -ep "请输入同步用的MySQL用户名（主从保持一致）: " master_user
+read -ep "请输入同步用的MySQL密码（主从保持一致）: " master_password
 read -ep "请输入MySQL服务器的ID（输入数字，注意不能重复，默认1为主服务器）: " mysql_id
 cat  >> /etc/mysql/mysql.conf.d/mysqld.cnf <<EOF 
 server-id = $mysql_id 
@@ -273,35 +369,42 @@ elif [[ -f /etc/debian_version ]]; then
 echo -e "# 检测到系统为Debian，仅支持MySQL8.0"
 # 检测MySQL安装包
 while true; do
-    if [ ! -f mysql*.tar ]; then
-        echo "需要的 mysql.tar 文件不存在，是否通过网络下载？(Y/N)"
+    if ! command -v mysql &> /dev/null; then
+        echo "未检测到MySQL环境，是否通过网络下载并安装？(Y/N)"
         read -r download_mysql
         if [ "$download_mysql" = "Y" ] || [ "$download_mysql" = "y" ]; then
-            wget -c -N --no-check-certificate https://downloads.mysql.com/archives/get/p/23/file/mysql-server_8.0.33-1debian11_amd64.deb-bundle.tar
-            break  # 如果下载完成，退出循环
+            echo "正在下载 MySQL 安装包..."
+            yum install -y wget && wget -c -N --no-check-certificate https://downloads.mysql.com/archives/get/p/23/file/mysql-server_8.0.33-1debian11_amd64.deb-bundle.tar
         else
             echo "请将 mysql.tar 文件放置到当前目录下，然后按回车键继续..."
             read  # 这里等待用户按回车键
+            if ls mysql*.tar 1> /dev/null 2>&1; then
+                echo "检测到MySQL安装包已存在。"
+                # 安装MySQL8.0依赖
+                wget -c -N --no-check-certificate https://mirrors.tuna.tsinghua.edu.cn/debian/pool/main/o/openssl/libssl1.1_1.1.1w-0%2Bdeb11u1_amd64.deb 
+                dpkg -i libssl1.1_1.1.1w-0+deb11u1_amd64.deb
+                # 解压MySQL安装包
+                tar -xvf mysql*.tar
+                # 安装MySQL
+                apt install -y psmisc libaio1 libnuma1 libatomic1 libmecab2 redis
+                dpkg -i mysql-common*.deb
+                dpkg -i mysql-community-client-plugins*.deb
+                dpkg -i mysql-community-client-core*.deb
+                dpkg -i mysql-community-client*.deb
+                dpkg -i mysql-client*.deb
+                dpkg -i mysql-community-server-core*.deb
+                dpkg -i mysql-community-server*.deb 
+                dpkg -i mysql-server*.deb
+                break  # 如果下载完成，退出循环
+            fi
         fi
     else
-        break  # 如果文件存在，退出循环
+        echo "检测到已有MySQL环境，跳过安装"
+        break  # 如果MySQL已安装，退出循环
     fi
 done
-# 安装MySQL8.0依赖
-wget -c -N --no-check-certificate https://mirrors.tuna.tsinghua.edu.cn/debian/pool/main/o/openssl/libssl1.1_1.1.1w-0%2Bdeb11u1_amd64.deb 
-dpkg -i libssl1.1_1.1.1w-0+deb11u1_amd64.deb
-# 解压MySQL安装包
-tar -xvf mysql*.tar
-# 安装MySQL
-apt install -y psmisc libaio1 libnuma1 libatomic1 libmecab2 redis
-dpkg -i mysql-common*.deb
-dpkg -i mysql-community-client-plugins*.deb
-dpkg -i mysql-community-client-core*.deb
-dpkg -i mysql-community-client*.deb
-dpkg -i mysql-client*.deb
-dpkg -i mysql-community-server-core*.deb
-dpkg -i mysql-community-server*.deb 
-dpkg -i mysql-server*.deb
+# 安装Redis
+apt install -y redis
 # 获取MySQL密码
 read -ep "请输入主MySQL的root密码: " root_password
 # 修改密码并允许所有IP访问
@@ -315,8 +418,8 @@ echo "Redis密码已修改为${root_password}"
 mysql -uroot -p"$root_password" -e "USE mysql; UPDATE user SET host='%' WHERE user='root'; FLUSH PRIVILEGES;"
 # 创建同步所需的MYSQL用户
 read -ep "请输入MySQL主服务器的地址: " master_host
-read -ep "请输入同步用的MySQL用户名: " master_user
-read -ep "请输入同步用的MySQL密码: " master_password
+read -ep "请输入同步用的MySQL用户名（主从保持一致）: " master_user
+read -ep "请输入同步用的MySQL密码（主从保持一致）: " master_password
 read -ep "请输入MySQL服务器的ID（输入数字，注意不能重复，默认1为主服务器）: " mysql_id
 cat  >> /etc/mysql/mysql.conf.d/mysqld.cnf <<EOF 
 server-id = $mysql_id 
@@ -379,4 +482,3 @@ else
     echo "不支持的操作系统"
     exit 1
 fi
-
